add documentation and other changes to src code

+++ src/_algorithm.jl
@@ -22,7 +22,7 @@ struct SnoptWECAlgorithm{AF, AB, F, I, B} <: OptAlgorithm
     checkgradients::B
 end
 SnoptWECAlgorithm() = SnoptWECAlgorithm(1e-7, [3.0, 2.6, 2.2, 1.8, 1.4, 1.0, 1.0], [1e-3, 1e-3, 1e-4, 1e-4, 1e-5, 1e-6, 1e-6], [false, false, false, false, false, false, true], Int(1e6), true)
-SnoptWECAlgorithm(a) = SnoptWECAlgorithm(a, [3.0, 2.6, 2.2, 1.8, 1.4, 1.0, 1.0], [1e-3, 1e-3, 1e-4, 1e-4, 1e-5, 1e-6, 1e-6], [false, false, false, false, false, false, true], Int(1e6), true)
+SnoptWECAlgorithm(a::Float64) = SnoptWECAlgorithm(a, [3.0, 2.6, 2.2, 1.8, 1.4, 1.0, 1.0], [1e-3, 1e-3, 1e-4, 1e-4, 1e-5, 1e-6, 1e-6], [false, false, false, false, false, false, true], Int(1e6), true)
 
 
 """
